% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_rej}
\alias{exp_rej}
\title{TUVN exponential rejection sampling}
\usage{
exp_rej(a, b)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncation}
}
\value{
value of the sample
}
\description{
TUVN exponential rejection sampling
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

