% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_b.R
\name{lower_b1_r}
\alias{lower_b1_r}
\title{lower bound for b - half-normal vs uniform}
\usage{
lower_b1_r(a)
}
\arguments{
\item{a}{Lower bound of truncated univariate normal}
}
\value{
value of the bound for b
}
\description{
Calculates the lower bound in 2.1.1.5 of Li and Ghosh. Called b_1(a) in the
    case breakdown
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

