% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_b.R
\name{lower_b2_r}
\alias{lower_b2_r}
\title{bound for b - exponential vs uniform sampling}
\usage{
lower_b2_r(a)
}
\arguments{
\item{a}{Lower bound of truncated univariate normal}
}
\value{
value of the bound for b
}
\description{
Calculates the bound for b in lemma 2.1.1.6 of Li & Ghosh. Called b_2(a)
    in the case breakdown.
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

