% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_b.R
\name{lower_b_r}
\alias{lower_b_r}
\title{lower bound for b - normal vs. uniform}
\usage{
lower_b_r(a)
}
\arguments{
\item{a}{Lower bound of truncated univariate normal}
}
\value{
value of the bound for b
}
\description{
Calculates the lower bound in Lemma 2.1.1.4 in Li & Ghosh (2015). It is the
    bound for b case 2.
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

