% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{norm_rej_r}
\alias{norm_rej_r}
\title{TUVN normal rejection sampling}
\usage{
norm_rej_r(a, b = Inf)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncation; default is infinity}
}
\value{
\item{x}{value of the sample} \item{acc}{number of proposals for acceptance}
}
\description{
Normal rejection sampling for the truncated standard normal distribution
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

