% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmvn.R
\name{rtmvn_r}
\alias{rtmvn_r}
\title{Truncated Multivariate Normal Distribution}
\usage{
rtmvn_r(n, Mean, Sigma = NULL, D, lower, upper, init, Sigma_chol = NULL)
}
\arguments{
\item{n}{number of samples to be generated}

\item{Mean}{mean vector}

\item{Sigma}{covariance matrix}

\item{D}{matrix of linear constraints}

\item{lower}{vector of lower bounds}

\item{upper}{vector of upper bounds}

\item{init}{vector of initial values for the Gibbs sampler. Must satisfy
the linear constraints.}

\item{Sigma_chol}{the lower triangular cholesky of the covariance matrix.
Only one of Sigma and Sigma_chol can be NULL.}
}
\value{
a matrix of samples with each column being an idependent sample.
}
\description{
Random vector generation for the truncated multivariate normal distribution
    using a Gibbs sampler.
}
\examples{
Mean = rep(0,2)
rho = 0.5
Sigma = matrix(c(10,rho,rho,0.1),2,2)
D = matrix(c(1,1,1,-1),2,2)
varp = Sigma[1,1]+Sigma[2,2]+2*Sigma[1,2] # var of the sum
varm = Sigma[1,1]+Sigma[2,2]-2*Sigma[1,2] # var of the diff
sd = c(sqrt(varp),sqrt(varm))
lower = -1.5*sd; upper = 1.5*sd; init = rep(0,2)
n = 20
rtmvn(n,Mean,Sigma,D,lower,upper,init)

}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

