% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtuvn}
\alias{rtuvn}
\title{Univariate Truncated Normal Distribution}
\usage{
rtuvn(n, mean, sd, lower, upper)
}
\arguments{
\item{n}{number of samples}

\item{mean}{mean}

\item{sd}{standard deviation}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\value{
a vector of generated samples
}
\description{
Random number generator for the truncated normal distribution. To
    calculate the density please use dtuvn.
}
\examples{
# sample from truncated normal with mean 10, sd 20, and lower = 10, upper = 20
rtuvn(n = 1, mean = 10, sd = 20, lower = 10, upper = 20)

}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

