% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_case1}
\alias{sample_case1}
\title{Sample Case 1}
\usage{
sample_case1(a, b)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncattion}
}
\value{
the value of the sample
}
\description{
generate a sample from case 1 where the bounds are \eqn{[a, \infty)}.
   The cases are listed on page 4 of Li & Ghosh (2015).
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

