% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_cases.R
\name{sample_case2_r}
\alias{sample_case2_r}
\title{Sample Case 2}
\usage{
sample_case2_r(a, b)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncattion}
}
\value{
\item{x}{value of the sample} \item{acc}{number of proposals for acceptance}
}
\description{
generate a sample from case 2 where the bounds are \eqn{0 \in [a, b]}.
   The cases are listed on page 4 of Li & Ghosh (2015).
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

