% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_case5}
\alias{sample_case5}
\title{Sample Case 5}
\usage{
sample_case5(a, b)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncattion}
}
\value{
value of the sample
}
\description{
generate a sample from case 5 where the bounds are \eqn{[a, b], b < 0}.
   This case is symmetric to case 3 in the function sample_case3.
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

