% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unif_rej}
\alias{unif_rej}
\title{TUVN uniform rejection sampling}
\usage{
unif_rej(a, b)
}
\arguments{
\item{a}{Lower bound of truncation}

\item{b}{Upper bound of truncation}
}
\value{
\item{x}{value of the sample} \item{acc}{number of proposals for acceptance}
}
\description{
TUVN uniform rejection sampling
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

