% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvn_gibbs}
\alias{rtmvn_gibbs}
\title{Gibbs sampler for the Truncated Multivariate Normal Distribution}
\usage{
rtmvn_gibbs(n, p, Mean, Sigma_chol, R, a, b, z)
}
\arguments{
\item{n}{number of samples to be generated}

\item{p}{dimension of the distribution to sample}

\item{R}{matrix of linear constraints}

\item{a}{vector of lower bounds}

\item{b}{vector of upper bounds}

\item{z}{vector of initial values for the Gibbs sampler.}
}
\value{
a matrix of samples with each column being an idependent sample.
}
\description{
Random vector generation for the truncated multivariate normal distribution
    using a Gibbs sampler.
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}

